fix: fix ls
