+++ packages/core/src/App.vue
@@ -1,10 +1,18 @@
 <script setup lang="ts">
 import { RouterView } from 'vue-router'
 import Loading from '@Com/Loading/index.vue'
-import { CustomLocalStorage } from 'utils'
-const ls = new CustomLocalStorage('SG', 60)
-// ls.setItem<string>('SG', '21', true)
-console.log(ls.getItem<string>('SG'))
+// import { Ls } from 'utils'
+// Ls('SG').setItem(
+//   'PG',
+//   {
+//     name: 'SG',
+//     age: 21,
+//   },
+//   false,
+//   1000
+// )
+// Ls('SG').setItem('M', 1, true, 1000)
+// console.log(Ls('SG').getItem('M'))
 </script>
 
 <template>
