+++ packages/utils/src/ls/index.ts
@@ -1,3 +1,1 @@
-export * from './core'
 export * from './ls'
-export * from './types'
